# back-end/docs/swagger.yaml
definitions:
...
  models.Room:
    type: object
  models.Seat:
    type: object
...
host: 10.177.88.190:8800
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: iBooking back-end api.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: iBooking
  version: "1.0"
paths:
...
  /room/:
    get:
      consumes:
      - application/json
      description: get all room information
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: get all room
      tags:
      - Room
  /room/auth/:
    get:
      consumes:
      - application/json
      description: get room by id
      parameters:
      - description: Create Room by giving room information
        in: path
        name: admin
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: get room
      tags:
      - Room
  /room/auth/createRoom:
    post:
      consumes:
      - application/json
      description: create room
      parameters:
      - description: Create Room by giving room information
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.Room'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: create room
      tags:
      - Room
  /room/auth/deleteRoom:
    post:
      consumes:
      - application/json
      description: delete room by id
      parameters:
      - description: Delete Room by giving room id
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: delete room
      tags:
      - Room
  /room/auth/updateRoom:
    post:
      consumes:
      - application/json
      description: update room
      parameters:
      - description: Update Room by giving room id and room details
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/models.Room'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: update room
      tags:
      - Room
  /seat/:
    get:
      description: get all seats
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: get seat
      tags:
      - Seat
  /seat/{seat_id}:
    get:
      consumes:
      - application/json
      description: get seat by id
      parameters:
      - description: Get Seat by giving seat id
        in: path
        name: seat_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: get seat by id
      tags:
      - Seat
  /seat/auth/createSeat:
    post:
      consumes:
      - application/json
      description: create seat
      parameters:
      - description: Create Seat by giving seat information
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.Seat'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: create seat
      tags:
      - Seat
  /seat/auth/deleteSeat:
    post:
      consumes:
      - application/json
      description: delete seat by id
      parameters:
      - description: Delete Seat by giving seat id
        in: body
        name: seat_id
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: delete seat
      tags:
      - Seat
  /seat/auth/updateSeat:
    post:
      consumes:
      - application/json
      description: update seat
      parameters:
      - description: Update Seat by giving seat information
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.Seat'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: update seat
      tags:
      - Seat
  /seat/getSeatByRoomID/{room_id}:
    get:
      consumes:
      - application/json
      description: get seat by room id
      parameters:
      - description: Get Seats by giving room id
        in: path
        name: room_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: get seat by room id
      tags:
      - Seat
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"